requirements.txt
yfinance>=0.2.18
pandas>=1.5.0
scikit-learn>=1.3.0
gspread>=5.10.0
oauth2client>=4.1.3
python-telegram-bot>=20.0
python-dotenv>=1.0.0
schedule>=1.2.0
numpy>=1.21.0,<2.0.0
requests>=2.28.0